+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 - Align attributes available in the instrumentation bus for listener related events.
 - Include consumer group id in consumption related events (#1093)
+- Remove unused logger listener event handler.
 
 ## 2.0.15 (2022-10-20)
 - Sanitize admin config prior to any admin action.